backend/requirements.txt
Flask>=2.0.0
flask-cors>=3.0.0
pymongo
werkzeug
pyjwt
